+++ tests/integration-network/src/mockchain/fpi/note/mod.rs
@@ -1,9 +1,12 @@
 //! Note-to-account foreign procedure invocation tests.
 
 mod four_words_struct;
+mod get_id_collision;
 mod mixed_scalar_struct;
 mod multi_call;
+mod multiple_components_active;
 mod multiple_packages;
+mod multiple_packages_aliased;
 mod no_arg;
 mod signed_narrow_indirect_struct;
 mod six_u64_struct;